Rename: Sollect -> Solelect로 변경 (#26)
