+++ src/main/java/com/ilta/solepli/domain/solelect/entity/Solelect.java
@@ -1,5 +1,9 @@
-package com.ilta.solepli.domain.sollect.entity.mapping;
+package com.ilta.solepli.domain.solelect.entity;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -7,6 +11,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -16,27 +21,32 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 
-import com.ilta.solepli.domain.place.entity.Place;
-import com.ilta.solepli.domain.sollect.entity.Sollect;
+import com.ilta.solepli.domain.solelect.entity.mapping.SolelectPlace;
+import com.ilta.solepli.domain.user.entity.User;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-@Table(name = "sollect_places")
-public class SollectPlace {
+@Table(name = "solelects")
+public class Solelect {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "sollect_id")
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  private Sollect sollect;
+  private String title;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "place_id")
+  @JoinColumn(name = "user_id")
   @OnDelete(action = OnDeleteAction.CASCADE)
-  private Place place;
+  private User user;
+
+  @OneToMany(mappedBy = "solelect", cascade = CascadeType.ALL, orphanRemoval = true)
+  @Builder.Default
+  private List<SolelectPlace> sollectPlaces = new ArrayList<>();
+
+  @OneToMany(mappedBy = "solelect", cascade = CascadeType.ALL, orphanRemoval = true)
+  @Builder.Default
+  private List<SolelectContent> solelectContents = new ArrayList<>();
 }
