+++ src/main/java/com/ilta/solepli/domain/solelect/entity/SolelectType.java
@@ -1,0 +1,6 @@
+package com.ilta.solepli.domain.solelect.entity;
+
+public enum SolelectType {
+  TEXT,
+  IMAGE;
+}
