+++ src/main/java/com/ilta/solepli/domain/solelect/entity/mapping/SolelectPlace.java
@@ -1,9 +1,6 @@
-package com.ilta.solepli.domain.sollect.entity;
+package com.ilta.solepli.domain.solelect.entity.mapping;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -19,30 +16,27 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 
+import com.ilta.solepli.domain.place.entity.Place;
+import com.ilta.solepli.domain.solelect.entity.Solelect;
+
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-@Table(name = "sollect_places")
-public class SollectContent {
+@Table(name = "solelect_places")
+public class SolelectPlace {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
-  @Column(nullable = false)
-  private Long seq;
-
-  @Enumerated(EnumType.STRING)
-  @Column(nullable = false)
-  private SollectType type;
-
-  private String text;
-
-  private String imageUrl;
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "solelect_id")
+  @OnDelete(action = OnDeleteAction.CASCADE)
+  private Solelect solelect;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "sollect_id")
+  @JoinColumn(name = "place_id")
   @OnDelete(action = OnDeleteAction.CASCADE)
-  private Sollect sollect;
+  private Place place;
 }
