--- src/main/java/com/ilta/solepli/domain/sollect/entity/Sollect.java
@@ -1,52 +1,0 @@
-package com.ilta.solepli.domain.sollect.entity;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-
-import com.ilta.solepli.domain.sollect.entity.mapping.SollectPlace;
-import com.ilta.solepli.domain.user.entity.User;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-@Table(name = "sollects")
-public class Sollect {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  private String title;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "user_id")
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  private User user;
-
-  @OneToMany(mappedBy = "sollect", cascade = CascadeType.ALL, orphanRemoval = true)
-  @Builder.Default
-  private List<SollectPlace> sollectPlaces = new ArrayList<>();
-
-  @OneToMany(mappedBy = "sollect", cascade = CascadeType.ALL, orphanRemoval = true)
-  @Builder.Default
-  private List<SollectContent> sollectContents = new ArrayList<>();
-}
