--- src/main/java/com/ilta/solepli/domain/sollect/entity/SollectType.java
@@ -1,6 +1,0 @@
-package com.ilta.solepli.domain.sollect.entity;
-
-public enum SollectType {
-  TEXT,
-  IMAGE;
-}
